GGUF: de-offset norms via keep-in-fp32 instead of pre-applying
SubtractOne is a real fp32 op again. Rather than pre-subtracting on the fp32
source in load_checkpoint_state, the quantizer forces the `w + 1` norms to stay
fp32 for the load via `_keep_in_fp32_modules_strict` (covers fp16 and bf16),
keyed off the SubtractOne converter targets in weight_mapping — no arch list.
The converter then subtracts in fp32 and the result is exact.

Those norms end up fp32 on the GGUF load, so the weights-conversion tests compare
at the original dtype. Validated on MPS: nemotron(fp16)/gemma2(fp16)/
gemma3_text(bf16) conversion + stablelm control + qwen2 generation.

+++ tests/quantization/ggml/test_ggml.py
@@ -841,7 +841,7 @@ class GgufModelTests(unittest.TestCase):
         for layer_name, original_params in original_state_dict.items():
             if layer_name in converted_state_dict:
                 self.assertTrue(original_params.shape == converted_state_dict[layer_name].shape)
-                torch.testing.assert_close(original_params, converted_state_dict[layer_name])
+                torch.testing.assert_close(original_params, converted_state_dict[layer_name].to(original_params.dtype))
             else:
                 raise ValueError(f"Layer {layer_name} is not presented in GGUF model")
 
@@ -919,7 +919,7 @@ class GgufModelTests(unittest.TestCase):
         for layer_name, original_params in original_state_dict.items():
             if layer_name in converted_state_dict:
                 self.assertTrue(original_params.shape == converted_state_dict[layer_name].shape)
-                torch.testing.assert_close(original_params, converted_state_dict[layer_name])
+                torch.testing.assert_close(original_params, converted_state_dict[layer_name].to(original_params.dtype))
             else:
                 raise ValueError(f"Layer {layer_name} is not presented in GGUF model")
 
@@ -958,7 +958,7 @@ class GgufModelTests(unittest.TestCase):
         for layer_name, original_params in original_state_dict.items():
             if layer_name in converted_state_dict:
                 self.assertTrue(original_params.shape == converted_state_dict[layer_name].shape)
-                torch.testing.assert_close(original_params, converted_state_dict[layer_name])
+                torch.testing.assert_close(original_params, converted_state_dict[layer_name].to(original_params.dtype))
             else:
                 raise ValueError(f"Layer {layer_name} is not presented in GGUF model")
 
@@ -983,7 +983,7 @@ class GgufModelTests(unittest.TestCase):
         for layer_name, original_params in original_state_dict.items():
             if layer_name in converted_state_dict:
                 self.assertTrue(original_params.shape == converted_state_dict[layer_name].shape)
-                torch.testing.assert_close(original_params, converted_state_dict[layer_name])
+                torch.testing.assert_close(original_params, converted_state_dict[layer_name].to(original_params.dtype))
             else:
                 raise ValueError(f"Layer {layer_name} is not presented in GGUF model")
